Refactor: Delete trash code
