--- project/frontend/js/mock/mockFriends.json
@@ -1,7 +1,0 @@
-[
-  {"friend_id": 1},
-  {"friend_id": 2},
-  {"friend_id": 3},
-  {"friend_id": 4},
-  {"friend_id": 5}
-]
