--- project/frontend/js/mock/mockProfile.json
@@ -1,11 +1,0 @@
-{
-  "avatar_name": "Wink Warrior",
-  "bio": "Lorem ipsum dolor, sit amet consectetur adipisicing elit. Animi voluptate minus ducimus placeat nemo delectus eveniet temporibus deleniti. Doloremque dolore distinctio enim nostrum illum ad aperiam porro placeat",
-  "email": "Wink_Warrior@42mail.com",
-  "total_games_play": 150,
-  "wins": 140,
-  "losses": 10,
-  "tour_play": 3,
-  "tour_won": 1,
-  "image": "../js/mock/wink.png"
-}
